message-sender/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyyaml==6.0.1
aiohttp==3.9.1
python-dotenv==1.0.0
